% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission_statistics.R
\name{admission_statistics}
\alias{admission_statistics}
\title{Summarises admissions}
\usage{
admission_statistics(groupingVariable = "studia", registrations = NULL,
  scores = NULL, exams = NULL, limits = NULL, output = NULL)
}
\arguments{
\item{groupingVariable}{optionally name of a grouping variable that should be
used in analysis given as a string or as a name/symbol}

\item{registrations}{optionally path to the file with data on registrations}

\item{exams}{optionally path to the file with data on examination scores}

\item{limits}{optionally path to the file with data on limits}

\item{output}{optionally path to the file in which results will be saved;
if \code{NA} is given as a value of this parameter, results won't be saved
to a file}
}
\value{
Data frame (tibble) with computed summary statistics - the same as
written to a file described with the \code{output} parameter (data frame is
returned invisibly).
}
\description{
The function computes a set of indicators describing admission process.
}
\details{
Location of files contaninig the data to be checked and location of file
in which results will be saved can be described noninteractively with
function arguments described above or - if any of this arguments is omitted -
interactively with a system file-selection dialog.
}
\examples{
\dontrun{
  admisssion_statistics()
}
}
