% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admission_statistics3.R
\name{admission_statistics3}
\alias{admission_statistics3}
\title{Summarises admissions without matura exams and minimal number of observations}
\usage{
admission_statistics3(groupingVariable = "studia",
  registrations = NULL, exams = NULL, limits = NULL, output = NULL)
}
\arguments{
\item{groupingVariable}{optionally name of a grouping variable that should be
used in analysis given as a string or as a name/symbol}

\item{registrations}{optionally path to a file with data on registrations}

\item{exams}{optionally path to a file with data on matura exams
if \code{NA} is given as a value of this parameter, results won't be saved
to a file}

\item{limits}{optionally path to a file with data on limits}

\item{output}{optionally path to a file in which results will be saved;}
}
\value{
Data frame (tibble) with computed summary statistics - the same as
written to a file described with the \code{output} parameter (data frame is
returned invisibly).
}
\description{
The function computes a set of indicators describing admission process. 
The function differs from admissions_statistics by focusing on the number of administrative 
processes instead of on the number of people. It does not use data on exams and does not check 
if there is at least 10 observations when raporting the results.
}
\details{
Location of files contaninig the data to be checked and location of file
in which results will be saved can be described noninteractively with
function arguments described above or - if any of this arguments is omitted -
interactively with a system file-selection dialog.
}
\examples{
\dontrun{
  admission_statistics2("studia", "inst/pr_pop_rek.csv",
                       "inst/as_egzaminy.xlsx", "inst/as_limity.xlsx",
                       "statistics.csv")
}
}
