% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exams.R
\name{check_exams}
\alias{check_exams}
\title{Checks if data on exams are complete and correct}
\usage{
check_exams(exams = exams)
}
\arguments{
\item{registrations}{optionally path to the file with data on exams}
}
\value{
if errors are detected: data frame with added columns indicating which observations should be fixed
if data are correct (corrected with the function) a data frame ready for further processing.
}
\description{
We should write something here
}
