% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_output_path.R
\name{check_output_path}
\alias{check_output_path}
\title{Checking correctness of parameter describing output file's location}
\usage{
check_output_path(file, paramName = "")
}
\arguments{
\item{file}{value of a parameter that should be checked}

\item{paramName}{optionally name of a parameter (to provide informative
error messages)}
}
\value{
Function returns \code{TRUE} if it doesn't detect problems. It
returns \code{FALSE} or \code{NA} if a file described by \code{file}
parameteralready exists and user choose not to overwrite it.
}
\description{
Simple routine developed to check correctness of parameters
that should desribe paths to output files
}
\details{
Function throws an error if it detects problem.
}
