% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_assert.R
\name{check_path}
\alias{check_path}
\title{Checking correctness of parameter describing file's location}
\usage{
check_path(file, param_name = "")
}
\arguments{
\item{file}{value of a parameter that should be checked}

\item{param_name}{optionally name of a parameter (to provide informative
error messages)}
}
\value{
Function returns nothing.
}
\description{
Simple routine developed to check correctness of parameters
that should desribe paths to files
}
\details{
Function throws an error if it detects problem.
}
