% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_rec_joining.R
\name{check_rec_joining}
\alias{check_rec_joining}
\title{Checks if there are codes in tha data that aren't in a recoding dictionary}
\usage{
check_rec_joining(data, rec, dataDescription = "pliku z danymi",
  recDescription = "pliku ze slownikiem")
}
\arguments{
\item{data}{data frame with data}

\item{rec}{data frame with dictionary}

\item{dataDescription}{desription of the data}

\item{recDescription}{description of the dictionary}
}
\value{
numeric value: 1 if user choose to remove rows that don't appear
in a dictionary (or no such rows were detected); 2 if user choose to use
this rows with codes as they appear in the data
}
\description{
Function checks if there are codes in tha data that aren't in
a recoding dictionary and if they are provides user with choice, how to deal
with them.
}
