% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_recregistrations.R
\name{check_recregistrations}
\alias{check_recregistrations}
\title{Function checks IRK recoding dictionary}
\usage{
check_recregistrations(recRegistrations = recRegistrations)
}
\arguments{
\item{recRegistrations}{data frame with recoding dictionary for
registrations data}
}
\value{
An error message if any variable is missing
}
\description{
Function checks if all required variables in IRK recoding
dictionary are present.
}
