% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_recusos.R
\name{check_recusos}
\alias{check_recusos}
\title{Function checks USOS recoding dictionary}
\usage{
check_recusos(recUsos)
}
\arguments{
\item{recUsos}{data frame with recoding dictionary for USOS data}
}
\value{
An error message if any variable is missing
}
\description{
Function checks if all required variables in USOS recoding
 dictionary are present.
}
