% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_registrations.R
\name{check_registrations}
\alias{check_registrations}
\title{Checks if data on registretions are complete and correct}
\usage{
check_registrations(registrations = registrations)
}
\arguments{
\item{registrations}{optionally path to the file with data on registrations}
}
\value{
if errors are detected: data frame with added columns indicating which observations should be fixed
if data are correct (corrected with the function) a data frame ready for further processing.
}
\description{
We should write something here
}
