% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_registrations_with_scores.R
\name{check_registrations_with_scores}
\alias{check_registrations_with_scores}
\title{Checks if data on registretions are complete and correct}
\usage{
check_registrations_with_scores(registrations)
}
\arguments{
\item{registrations}{data frame with joined data regarding registrations and
regarding recruitment scores}
}
\value{
data frame
}
\description{
Function checks if data on registrations and data on recruitment
scores joined together are correct. Function gives user a choice whether to
correct detected minor issues detected in the data, or to add column
indicating which observations should be fixed. Finally it returns a data
frame with corrected data or columns indicating problems.
}
