% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_results.R
\name{check_scores}
\alias{check_scores}
\title{Checks if data on scores are complete and correct}
\usage{
check_scores(scores = scores)
}
\arguments{
\item{registrations}{optionally path to the file with data on scores}
}
\value{
if errors are detected: data frame with added columns indicating which observations should be fixed
if data are correct (corrected with the function) a data frame ready for further processing.
}
\description{
We should write something here
}
