% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variable_names.R
\name{check_variable_names}
\alias{check_variable_names}
\title{Function checks IRK recoding dictionary}
\usage{
check_variable_names(x, requiredNames, xDescription = "ramce danych")
}
\arguments{
\item{x}{data frame to be checked}

\item{requiredNames}{character vector containing names of variables that are
required in a checked data frame}

\item{xDescription}{optionally a string describing \code{x} in a user-friendly
way, used to construct an error message that is thrown thrown if some of
required variables are not found}
}
\value{
An error message if any variable is missing
}
\description{
Function checks if all required variables in IRK recoding
dictionary are present.
}
