% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variable_values.R
\name{check_variable_values}
\alias{check_variable_values}
\title{Checks if a variable has values within a defined set}
\usage{
check_variable_values(x, valMin = NULL, valMax = NULL, valSet = NULL)
}
\arguments{
\item{x}{a variable to be checked}

\item{valMin}{a minimal allowed value, an optional parameter}

\item{valMax}{a maximal allowed value, an optional parameter}

\item{valSet}{a set of allowed values, an optional parameter}
}
\value{
if errors are detected: a warrning message listing errors
and a menu offering an option to stop the program
}
\description{
Function checks if a variable has values within a defined set
and checks the number of NAs.
}
