% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_file.R
\name{choose_file}
\alias{choose_file}
\title{Selecting file to read}
\usage{
choose_file(about = "", errorOnCancel = TRUE)
}
\arguments{
\item{about}{optionally a string describing what should contain a file that
user is supposed to select}

\item{errorOnCancel}{logical value (TRUE or FALSE) - should canceling
a file selection dialog should throw an error?}
}
\value{
String containing a path to a file.
}
\description{
Simple routine regarding choosing interactively files to read
}
\details{
If no file is selected, function throws an error (parameter
\code{about} can help to make this error more informative).
}
