% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foreigners.R
\name{foreigners}
\alias{foreigners}
\title{Lists foreigners and their admissions status}
\usage{
foreigners(registrations = NULL, scores = NULL, output = NULL)
}
\arguments{
\item{registrations}{optionally path to the file with data on registrations}

\item{scores}{optionally path to the file with data on recruitment scores}

\item{output}{optionally path to the file in which results will be saved;
if \code{NA} is given as a value of this parameter, results won't be saved
to a file}
}
\value{
Data frame (tibble) with the list of foreign applicants - at the same as
written to a file described with the \code{output} parameter (data frame is
returned invisibly).
}
\description{
We should write something here
}
\details{
Location of files contaninig the data to be checked and location of file
in which results will be saved can be described noninteractively with
function arguments described above or - if any of this arguments is omitted -
interactively with a system file-selection dialog.
}
