% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_registrations.R
\name{prepare_registrations}
\alias{prepare_registrations}
\title{Preparing recrutation data for analysis}
\usage{
prepare_registrations(registrations = NULL, scores = NULL,
  output = NULL, usosAdmission = NULL, recRegistrations = NULL,
  recUsos = NULL)
}
\arguments{
\item{registrations}{optionally a path to the file with data on registrations}

\item{scores}{optionally a path to the file with data on recruitment scores}

\item{output}{optionally a path to the file in which results will be saved}

\item{usosAdmission}{optionally a path to the file with USOS data on admissions}

\item{recRegistrations}{optionally a path to the file with instructions to
alter programme codes in IRK data}

\item{recUsos}{optionally a path to the file with instructions to alter
programme codes in USOS data}
}
\value{
\itemize{
  \item{If problems were found in registrations data or recruitment scores
  data and user choose to tag (at least some of) them, function returns
  invisibly a list with two elements that are data frames:
  \code{registrations} and \code{scores}. These are data frames containing
  input data with tagged problems (and possibly corrected other problems,
   that user choose not to tag, but to correct).}
  \item{In any other case function returns invisibly a data frame with
  corrected data on registrations.}
}
}
\description{
Function creates a corrected dataset on registrations that is
suitable for final analysis. Operations: checking against USOS records on
acepted applicants, checking and correcting multiple registrations i.e. cases
that an applicant applies more than one time to a given program, adding data
on scores to the data on registrations.
}
\examples{
\dontrun{
prepare_registrations("inst/pr_zapisy.xlsx", "inst/pr_wyniki.xlsx",
                      "dane/dane-IRK poprawione.csv", "inst/pr_usos.xlsx",
                      "inst/pr_rek_irk.xlsx", "inst/pr_rek_usos.xlsx")
}
}
