% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_registrations3.R
\name{prepare_registrations3}
\alias{prepare_registrations3}
\title{Preparing recrutation data for analysis - a new test version}
\usage{
prepare_registrations3(registrations = NULL, scores = NULL,
  output = NULL, usosAdmission = NULL, recRegistrations = NULL,
  recUsos = NULL)
}
\arguments{
\item{registrations}{optionally a path to the file with data on registrations}

\item{scores}{optionally a path to the file with data on recruitment scores}

\item{output}{optionally a path to the file in which results will be saved}

\item{usosAdmission}{optionally a path to the file with USOS data on admissions}

\item{recRegistrations}{optionally a path to the file with instructions to
alter programme codes in IRK data}

\item{recUsos}{optionally a path to the file with instructions to alter
programme codes in USOS data}
}
\value{
\itemize{
  \item{If problems were found in registrations data or recruitment scores
  data and user choose to tag (at least some of) them, function returns
  invisibly a data frame with tagged (and possibly corrected with respect to
  other problems, that user chose to correct, not to tag) dataset.}
  \item{In any other case function returns invisibly a data frame with
  corrected data on registrations.}
}
}
\description{
Function creates a corrected dataset on registrations that is
suitable for final analysis. Operations: checking against USOS records on
acepted applicants, checking and correcting multiple registrations i.e. cases
that an applicant applies more than one time to a given program, adding data
on scores to the data on registrations. Creates two versions of scores: a max score as an applicant
and a max score as an acepted applicant. This might needed in the case of multiple paths
in the admissions process (e.g. an extra path for transfer students).
}
\examples{
\dontrun{
prepare_registrations3("inst/pr_zapisy.xlsx", "inst/pr_wyniki.xlsx",
                      "dane/dane-IRK poprawione.csv", "inst/pr_usos.xlsx",
                      "inst/pr_rek_irk.xlsx", "inst/pr_rek_usos.xlsx")
}
}
