% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file}
\alias{read_file}
\title{Reading file}
\usage{
read_file(file)
}
\arguments{
\item{file}{path to a file}
}
\value{
Data frame with a data that has been read from a file.
}
\description{
Routine for reading files
}
\details{
Only .xlsx or .csv files are permitted. CSV files can be comma,
semicolon or tab separated and should be UTF-8 encoded.
}
