% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_data.R
\name{summarise_data}
\alias{summarise_data}
\title{Performing summarising computations}
\usage{
summarise_data(x, g, f = TRUE, joinTo = NULL, ...)
}
\arguments{
\item{x}{data (to be summarised)}

\item{g}{grouping variable provided as a name (symbol)}

\item{f}{optionally a call defining filtering to be done before computations}

\item{joinTo}{optionally a data frame to which computed summary statistics
should be (left-) joined (if defined, function will return a joined object)}

\item{...}{calls describing computations to be performed}
}
\value{
Tibble.
}
\description{
This function is designed as a mid-level interface to perform
summarising computations. It requaries to define grouping (with the \code{g}
parameter, which should be given as a name/symbol) and allows to define
filtering, that will be done before performing computations (with the \code{f}
parameter, which should be given as a call). Computations to be performed
should be given in a dplyr style, that is as calls.
}
\details{
In a function call as a parameter \code{g} typically will be given
an object returned by \code{\link[rlang:quotation]{ensym}} or
\code{\link[rlang:quotation]{enexpr}} functions.
}
