% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarising_template.R
\name{summarising_template}
\alias{summarising_template}
\title{Template of function summarising the data}
\usage{
summarising_template(x, g)
}
\arguments{
\item{x}{data (to be summarised)}

\item{g}{grouping variable provided as a name (symbol)}
}
\value{
Tibble.
}
\description{
This function is a template showing the general structure of
a function designed to compute summary statistics on the analysed data
}
\details{
In a function call as a parameter \code{g} typically will be given
an object returned by \code{\link[rlang]{ensym}} or
\code{\link[rlang]{enexpr}} functions.
}
