% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Checking the correctness of a recrutation data}
\usage{
check_data(registrations = NULL, scores = NULL, exams = NULL)
}
\arguments{
\item{registrations}{optionally path to the file with data on registrations}

\item{scores}{optionally path to the file with data on recruitment scores}

\item{exams}{optionally path to the file with data on examination scores}
}
\value{
Function retruns nothing.
}
\description{
We should write somethin here
}
\details{
Location of files contaninig the data to be checked can be described
noninteractively with function arguments described above or - if any of this
arguments is omitted - interactively with a system file-selection dialog.
}
\examples{
\dontrun{
  check_data()
}
}
