% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Checking the correctness of a recrutation data}
\usage{
check_data(groupingVariable, registrations = NULL, scores = NULL,
  exams = NULL, dictionary = NULL, output = NULL,
  baseGroupingVariable = "studia")
}
\arguments{
\item{groupingVariable}{optionally name of a grouping variable that should be
used in analysis given as a string or as a name/symbol}

\item{registrations}{optionally path to the file with data on registrations}

\item{scores}{optionally path to the file with data on recruitment scores}

\item{exams}{optionally path to the file with data on examination scores}

\item{dictionary}{optionally path to the file with data containing additional
informations about registrations}

\item{output}{optionally path to the file in which results will be saved;
if \code{NA} is given as a value of this parameter, results won't be saved
to a file}

\item{baseGroupingVariable}{name of (to be described) variable given as
a string or as a name/symbol}
}
\value{
Data frame (tibble) with computed summary statistics - the same as
written to a file described with the \code{output} parameter (data frame is
returned invisibly).
}
\description{
We should write something here
}
\details{
Location of files contaninig the data to be checked and location of file
in which results will be saved can be described noninteractively with
function arguments described above or - if any of this arguments is omitted -
interactively with a system file-selection dialog.
}
\examples{
\dontrun{
  check_data()
}
}
