% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_exams.R
\name{check_exams}
\alias{check_exams}
\title{Checks if data on exams are complete and correct}
\usage{
check_exams(exams = exams)
}
\arguments{
\item{exams}{data frame with data regarding exams}
}
\value{
data frame
}
\description{
Function checks if data on exams are complete and
correct. If no critical eroors are detected, functions gives user a choice
whether to correct detected minor issues detected in the data, or to add
columns indicating which observations should be fixed. Finally it returns
a data frame with corrected data or columns indicating problems.
}
