% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_registrations.R
\name{check_registrations}
\alias{check_registrations}
\title{Checks if data on registretions are complete and correct}
\usage{
check_registrations(registrations)
}
\arguments{
\item{registrations}{data frame with data regarding registrations}
}
\value{
data frame
}
\description{
Function checks if data on registretions are complete and
correct. If no critical eroors are detected, functions gives user a choice
whether to correct detected minor issues detected in the data, or to add
columns indicating which observations should be fixed. Finally it returns
a data frame with corrected data or columns indicating problems.
}
