% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_with_check.R
\name{join_with_check}
\alias{join_with_check}
\title{Joining two data frames}
\usage{
join_with_check(x, y, xDescription = "x", yDescription = "y")
}
\arguments{
\item{x}{data frame}

\item{y}{data frame}

\item{xDescription}{optionally a description of what is \code{x} (for
purpose of providing informative warninig message)}

\item{yDescription}{optionally a description of what is \code{y}  (for
purpose of providing informative warninig message)}
}
\value{
Data frame returned by \code{full_join(x, y)}.
}
\description{
Simple routine designed to find out (and report) whether two
data frames have unique combinations of values of common variables and
whether all rows in \code{x} has pair(s) in \code{y} and \emph{vice versa}
before performing full join.
}
